requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.1.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
py-cord==2.4.1
python-dotenv==1.0.0
requests==2.28.1
typing_extensions==4.12.2
urllib3==1.26.20
yarl==1.17.2
yt-dlp==2024.11.18
